Create 1:N

+++ accounts/views.py
@@ -1,16 +1,16 @@
 from django.shortcuts import render, redirect
-from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
-from django.contrib.auth import login as auth_login, logout as auth_logout
-from django.views.decorators.http import require_POST
+from django.views.decorators.http import require_GET, require_POST
+from django.contrib.auth.forms import UserCreationForm, AuthenticationForm, PasswordChangeForm
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import login as auth_login, logout as auth_logout, update_session_auth_hash
+from .forms import CustomUserChangeForm
+
 # Create your views here.
-
-
 def signup(request):
     if request.user.is_authenticated:
         return redirect('articles:index')
 
     if request.method == 'POST':
-        #회원가입 로직
         form = UserCreationForm(request.POST)
         if form.is_valid():
             form.save()
@@ -18,23 +18,24 @@ def signup(request):
     else:
         form = UserCreationForm()
     context = {'form': form}
-    return render(request, 'accounts/signup.html', context)
+    return render(request, 'accounts/form.html', context)
+
 
 def login(request):
     if request.user.is_authenticated:
-        return redirect('articles:index')
+        return redirect('articles:index')    
 
     if request.method == 'POST':
         form = AuthenticationForm(request, request.POST)
         if form.is_valid():
-            #Session 만들어줘야함
             auth_login(request, form.get_user())
             next_page = request.GET.get('next')
             return redirect(next_page or 'articles:index')
     else:
         form = AuthenticationForm()
     context = {'form': form}
-    return render(request, 'accounts/login.html', context)
+    return render(request, 'accounts/form.html', context)
+
 
 def logout(request):
     auth_logout(request)
@@ -43,5 +44,34 @@ def logout(request):
 
 @require_POST
 def delete(request):
-    request.user.delete()
+    if request.user.is_authenticated:
+        request.user.delete()
     return redirect('articles:index')
+
+
+@login_required
+def update(request):
+    if request.method == 'POST':
+        form = CustomUserChangeForm(request.POST, instance=request.user)
+        if form.is_valid():
+            form.save()
+            return redirect('articles:index')
+    else:
+        form = CustomUserChangeForm(instance=request.user)
+    context = {'form': form}
+    return render(request, 'accounts/form.html', context)
+
+
+@login_required
+def password(request):
+    if request.method == 'POST':
+        form = PasswordChangeForm(request.user, request.POST)
+        if form.is_valid():
+            user = form.save()
+            update_session_auth_hash(request, user)
+            return redirect('accounts:update')
+    else:
+        form = PasswordChangeForm(request.user)
+    context = {'form': form}
+    return render(request, 'accounts/form.html', context)
+
